Test linter against copy of master branch code
The branch linter_test has iteratively linted code at this point, so it's hard to tell how the linter will affect the master branch.
When the linter runs for the first time on this new branch, it will demonstrate how code in the master branch will be changed upon merge.